+++ lib/src/xml/mixins/has_children.dart
@@ -1,12 +1,17 @@
 library xml.mixins.has_children;
 
+import '../nodes/element.dart';
 import '../nodes/node.dart';
+import '../utils/name_matcher.dart';
 import '../utils/node_list.dart';
 
 /// Children interface for nodes.
 mixin XmlChildrenBase {
   /// Return the direct children of this node in document order.
   List<XmlNode> get children => const [];
+
+  /// Return the first child element with the given `name`, or `null`.
+  XmlElement getElement(String name, {String namespace}) => null;
 
   /// Return the first child of this node, or `null` if there are no children.
   XmlNode get firstChild => null;
@@ -19,6 +24,11 @@ mixin XmlChildrenBase {
 mixin XmlHasChildren implements XmlChildrenBase {
   @override
   final XmlNodeList<XmlNode> children = XmlNodeList<XmlNode>();
+
+  @override
+  XmlElement getElement(String name, {String namespace}) => children
+      .whereType<XmlElement>()
+      .firstWhere(createNameMatcher(name, namespace), orElse: () => null);
 
   @override
   XmlNode get firstChild => children.isEmpty ? null : children.first;